requirements.txt
numpy
scipy
pandas
matplotlib
scikit-learn
pytorch-lightning
torchmetrics
einops
open3d
trimesh
pyrender
pyquaternion
smplx==0.1.28
vedo
vtk
plyfile
pycocotools
imageio
imageio-ffmpeg
opencv-python
tqdm
tensorboardX
plotly
line_profiler
tabulate
yacs
omegaconf
easydict
ConfigArgParse
gdown
requests
huggingface-hub
yapf==0.30.0
scikit-image
numba
cython
chumpy
Pillow
json_tricks
torchgeometry
boto3
requests
easydict
timm
pyglet
mmdet==2.26.0

-e git+https://github.com/open-mmlab/mmhuman3d.git@8c95747175f8c4bd76d5e0be7be244f8a9a4a6de#egg=mmhuman3d
# -e git+https://github.com/caizhongang/SMPLer-X.git@73cd8012eb22d7e46bcdaabff4bdc1d7dab607a9#egg=mmpose&subdirectory=main/transformer_utils
